Add custom exception classes instead of using plain error types.
This adds custom exception classes so that users do not have to match on
generic exception types in code that uses this module.

The new custom classes are structured so that existing code that is
matching on the generic exception types will still work, and so that
there's a primary exception type that can be matched on to catch all
errors specific to this module.